+++ CHANGELOG.md
@@ -13,3 +13,21 @@ the following form: YYYY.0M.0D.
 - Designed a `Triangle` component
 - Designed a `Data Table` component
 - Designed a `Projectile` component
+
+## [2026.02.26]
+
+### [Added]
+
+- Designed a proof of concept for `Triangle` component
+- Wrote description of intent for component proof of concept in `02-component-proof-of-concept.md`
+
+- Implemented constructor and `createNewRep(int dimensions)`
+- Implemented `addVertice(double[] coordingates)`
+- Implemented `removeAnyVertice()`
+- Implemented `area()`
+- Wrote a working demo of `area()` in the `main()` function
+
+### [Updated]
+
+- Minor changes to `01-compoent-brainstorming` by updating goal of `triangle` component
+    to ideally function in any number of dimensions greater than or equal to 2.